requirements.txt
PyQt6==6.9.1
ffmpeg-python==0.2.0